business-logic/requirements.txt
Flask
flask-cors
psycopg2-binary
PyJWT
prometheus_client